% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myEigen_QR.R
\name{myEigen_QR}
\alias{myEigen_QR}
\title{PCA}
\usage{
myEigen_QR(A, numIter = 1000)
}
\arguments{
\item{A}{variance covariance matrix}

\item{numIter}{number of iteration, default to 1000}
}
\description{
This function perform PCA on matrix A based on eigen decomposition with power method, return a list of the eigen value vector D and the matrix of eigen vectors V.
}
\examples{
M <- matrix(rnorm(25),nrow=5);
A <- t(M) \%*\% M
myEigen_QR(A)
}
