% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myLasso.R
\name{myLasso}
\alias{myLasso}
\title{Lasso regression}
\usage{
myLasso(X, Y, lambda_all)
}
\arguments{
\item{X}{design matrix}

\item{Y}{response vector}

\item{lambda_all}{vector of regularization parameters}
}
\description{
This function finds lasso solution path for various values of regularization parameter of Y regressed on X, return a matrix with each column corresponding to a regularization parameters.
}
\examples{
n <- 50
p <- 25
X <- matrix(rnorm(n * p), nrow = n)
beta <- c(rep(0,20),rep(1,5))
Y <- X \%*\% beta+rnorm(n)
lambda_all <- (100:1)/10
lasso <- myLasso(X,Y,lambda_all)
beta_sum <- t(matrix(rep(1, (p+1)), nrow = 1)\%*\%abs(beta_all))
matplot(beta_sum, t(beta_all), type = 'l', lty = 1,xlab="|beta|",ylab="beta")
text(max(beta_sum),beta_all[,length(lambda_all)],0:p,cex=1,col=1:p)
}
