% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myLM.R
\name{myLM}
\alias{myLM}
\title{Linear regression}
\usage{
myLM(X, Y, intercept = TRUE)
}
\arguments{
\item{X}{design matrix}

\item{Y}{response vector}

\item{intercept}{whether the model contains an intercept, default to TRUE}
}
\description{
This function perform linear regression of Y on X powered by QR decomposition, return a list of the least squares solution vector.
}
\examples{
n <- 100
p <- 4
X <- matrix(rnorm(n * p), nrow = n)
beta <- rnorm(p)
Y <- X \%*\% beta+rnorm(n)
myLM(X, Y, intercept=FALSE)
}
