% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myLogistic.R
\name{myLogistic}
\alias{myLogistic}
\title{Logistic regression}
\usage{
myLogistic(X, Y)
}
\arguments{
\item{X}{design matrix}

\item{Y}{response vector}
}
\description{
This function perform logistic regression of Y on X with IRLS method, return the IRLS solution vector.
}
\examples{
n <- 100
p <- 4
X <- matrix(rnorm(n * p), nrow = n)
beta <- rnorm(p)
expit <- function(x){
      1 / (1 + exp(-x))
}
Y <- 1 * (runif(n) < expit(X \%*\% beta))
myLogistic(X, Y)
}
