% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myQR.R
\name{myQR}
\alias{myQR}
\title{QR decomposition}
\usage{
myQR(A)
}
\arguments{
\item{A}{input matrix}
}
\description{
This function perform QR decomposition on the matrix A, return a list of orthogonal matrix Q and upper triangular matrix R
}
\examples{
A <- matrix(rnorm(25),nrow=5);
myQR(A)
}
