% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myRidge.R
\name{myRidge}
\alias{myRidge}
\title{Ridge regression}
\usage{
myRidge(X, Y, lambda)
}
\arguments{
\item{X}{design matrix}

\item{Y}{response vector}

\item{lambda}{regularization parameter (lambda >= 0)}
}
\description{
This function perform ridge regression of Y on X with QR method, return the coefficient vector.
}
\examples{
n <- 100
p <- 4
X <- matrix(rnorm(n * p), nrow = n)
beta <- rnorm(p)
Y <- X \%*\% beta+rnorm(n)
lambda=1
myRidge(X, Y, lambda)
}
