% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mySpline.R
\name{mySpline}
\alias{mySpline}
\title{Spline regression}
\usage{
mySpline(x, Y, lambda, p = 100)
}
\arguments{
\item{x}{n x 1 vector or matrix of explanatory variables}

\item{Y}{n x 1 response vector}

\item{lambda}{regularization parameter (lambda >= 0)}

\item{p}{number of knots to make to the x-axis}
}
\description{
This function perform one-dimensional spline regression of Y on x with ridge regularization, return a list containing the spline regression coefficient vector and the predicted Y values.
}
\examples{
n = 20
p = 100
sigma = 0.1
lambda = 1
x = runif(n)
x = sort(x)
x = matrix(x, nrow=n)
Y = x^2 + rnorm(n)*sigma
}
