% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mySweep.R
\name{mySweep}
\alias{mySweep}
\title{Sweep operation}
\usage{
mySweep(A, m)
}
\arguments{
\item{A}{a square matrix}

\item{m}{the pivot element is A[m, m]}
}
\description{
This function perform a SWEEP operation on A with the pivot element A[m,m]
}
